Adding turn updating for initial testing

+++ CNC/turn.py
@@ -337,7 +337,7 @@ class Turn:
             pol_stab_loss = ((user['last_pol_auth_gain']*2)-10) if user['last_pol_auth_gain'] < 5 else 0
             mil_stab_loss = ((user['last_mil_auth_gain']*2)-8) if user['last_mil_auth_gain'] > 4 else 0
             # stability gain from "strong army"
-            manpower_usage_count = await conn.fetchval('''SELECT SUM(troops) FROM cnc_armies 
+            manpower_usage_count = await conn.fetchval('''SELECT SUM(troops)::INT FROM cnc_armies 
                                                           WHERE owner_id = $1;''', user['user_id'])
             # protect against null
             manpower_usage_count = 0 if manpower_usage_count is None else manpower_usage_count
